Sector26

+++ ptychodus/model/image.py
@@ -8,10 +8,10 @@ import matplotlib.pyplot
 import numpy
 import numpy.typing
 
+from ..api.geometry import Interval
 from ..api.image import ScalarTransformation, ComplexToRealStrategy
 from ..api.observer import Observable, Observer
 from ..api.plugins import PluginChooser, PluginEntry
-from .geometry import Interval
 
 logger = logging.getLogger(__name__)
 
@@ -190,7 +190,7 @@ class ImagePresenter(Observable, Observer):
             if numpy.iscomplexobj(self._array):
                 self._image = complexToRealStrategy(self._array, scalarTransform)
             else:
-                self._image = scalarTransform(self._array)
+                self._image = scalarTransform(self._array.astype(numpy.float32))
 
         if self._image is None or numpy.size(self._image) <= 0:
             self._dataRange = Interval[Decimal](Decimal(0), Decimal(1))
